+++ src/hard/app/routes/exercises.py
